// src/domain/atributo/atributo.interface.ts
export class AtributoInterface
{
    forca: number
    destreza: number
    constituicao: number
    inteligencia: number
    sabedoria: number
    carisma: number
    personagemId: number
    forca_modificador?: number
    destreza_modificador?: number
    constituicao_modificador?: number
    sabedoria_modificador?: number
    inteligencia_modificador?: number
    carisma_modificador?: number
}

export class AtributoInterfaceUpdate
{
    forca?: number
    destreza?: number
    constituicao?: number
    inteligencia?: number
    sabedoria?: number
    carisma?: number
    personagemId?: number
    forca_modificador?: number
    destreza_modificador?: number
    constituicao_modificador?: number
    sabedoria_modificador?: number
    inteligencia_modificador?: number
    carisma_modificador?: number
}